move core interfaces to `commonMain`
