--- shared/src/androidMain/kotlin/com/github/odaridavid/weatherapp/Platform.android.kt
@@ -1,7 +1,0 @@
-package com.github.odaridavid.weatherapp
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
