--- shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/Greeting.kt
@@ -1,9 +1,0 @@
-package com.github.odaridavid.weatherapp
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
