--- shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/Platform.kt
@@ -1,7 +1,0 @@
-package com.github.odaridavid.weatherapp
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
