+++ shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/api/Logger.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.core.api
+package com.github.odaridavid.weatherapp.api
 
 interface Logger {
     fun logException(throwable: Throwable)
